refactor(workers): extract helpers to fix max-lines lint
