+++ apps/pops-api/src/modules/cerebrum/workers/consolidator.ts
@@ -1,3 +1,17 @@
+/**
+ * Consolidator Worker (US-02, PRD-085).
+ *
+ * Detects clusters of 3+ engrams with cosine similarity > 0.85 within the
+ * same top-level scope. Produces merge plans that combine content, deduplicate
+ * tags/links, and archive originals. Large clusters (>10) are split.
+ */
+import {
+  bfsComponent,
+  buildMergePlan,
+  chunk,
+  findDominantType,
+  groupByTopLevelScope,
+} from './consolidator-helpers.js';
 import {
   DEFAULT_CONSOLIDATOR_CONFIG,
   type ConsolidatePayload,
@@ -6,49 +20,12 @@ import {
   type GliaActionType,
   type WorkerRunResult,
 } from './types.js';
-/**
- * Consolidator Worker (US-02, PRD-085).
- *
- * Detects clusters of 3+ engrams with cosine similarity > 0.85 within the
- * same top-level scope. Produces merge plans that combine content, deduplicate
- * tags/links, and archive originals. Large clusters (>10) are split.
- */
-import { WorkerBase, topLevelScope, type WorkerBaseDeps } from './worker-base.js';
+import { WorkerBase, type WorkerBaseDeps } from './worker-base.js';
 
 import type { Engram } from '../engrams/types.js';
 
 export interface ConsolidatorDeps extends WorkerBaseDeps {
   config?: Partial<ConsolidatorConfig>;
-}
-
-/** Group engrams by their top-level scope for scope-isolated clustering. */
-function groupByTopLevelScope(engrams: Engram[]): Map<string, Engram[]> {
-  const groups = new Map<string, Engram[]>();
-  for (const engram of engrams) {
-    for (const scope of engram.scopes) {
-      const top = topLevelScope(scope);
-      const existing = groups.get(top) ?? [];
-      if (!existing.some((e) => e.id === engram.id)) {
-        existing.push(engram);
-        groups.set(top, existing);
-      }
-    }
-  }
-  return groups;
-}
-
-/** Deduplicate a string array. */
-function dedupe(values: string[]): string[] {
-  return [...new Set(values)];
-}
-
-/** Split an array into chunks of at most `size`. */
-function chunk<T>(arr: T[], size: number): T[][] {
-  const result: T[][] = [];
-  for (let i = 0; i < arr.length; i += size) {
-    result.push(arr.slice(i, i + size));
-  }
-  return result;
 }
 
 export class ConsolidatorWorker extends WorkerBase {
@@ -128,39 +105,14 @@ export class ConsolidatorWorker extends WorkerBase {
 
     for (const engram of engrams) {
       if (visited.has(engram.id)) continue;
-      const component = this.bfsComponent(engram.id, adjacency, visited, engramById);
+      const component = bfsComponent(engram.id, adjacency, visited, engramById);
       if (component.length >= minSize) clusters.push(component);
     }
     return clusters;
   }
 
-  private bfsComponent(
-    startId: string,
-    adjacency: Map<string, Set<string>>,
-    visited: Set<string>,
-    engramById: Map<string, Engram>
-  ): Engram[] {
-    const component: Engram[] = [];
-    const queue = [startId];
-
-    while (queue.length > 0) {
-      const current = queue.shift();
-      if (current === undefined || visited.has(current)) continue;
-      visited.add(current);
-
-      const e = engramById.get(current);
-      if (e) component.push(e);
-
-      const neighbours = adjacency.get(current) ?? new Set<string>();
-      for (const neighbour of neighbours) {
-        if (!visited.has(neighbour)) queue.push(neighbour);
-      }
-    }
-    return component;
-  }
-
   private processCluster(cluster: Engram[], scope: string, phase: string): GliaAction {
-    const mergePlan = this.buildMergePlan(cluster, scope);
+    const mergePlan = buildMergePlan(cluster, scope, this.engramService);
     const action = this.createAction(
       cluster.map((e) => e.id),
       `Consolidating ${cluster.length} similar engrams in scope '${scope}': ${cluster.map((e) => `"${e.title}"`).join(', ')}`,
@@ -174,44 +126,9 @@ export class ConsolidatorWorker extends WorkerBase {
     return action;
   }
 
-  /** Build a merge plan from a cluster of similar engrams. */
-  private buildMergePlan(cluster: Engram[], scope: string): ConsolidatePayload {
-    const allTags = dedupe(cluster.flatMap((e) => e.tags));
-    const clusterIds = new Set(cluster.map((e) => e.id));
-    const allLinks = dedupe(
-      cluster.flatMap((e) => e.links).filter((link) => !clusterIds.has(link))
-    );
-    const mergedTitle = `Consolidated: ${cluster[0]?.title ?? 'Untitled'}${cluster.length > 1 ? ` (+${cluster.length - 1} related)` : ''}`;
-
-    const bodyParts = cluster.map((e) => {
-      const { body } = this.engramService.read(e.id);
-      return `### From: ${e.title} (${e.id})\n\n${body}`;
-    });
-
-    const mergedBody = [
-      `# ${mergedTitle}`,
-      '',
-      ...bodyParts,
-      '',
-      '## Sources',
-      '',
-      ...cluster.map((e) => `- ${e.id}: ${e.title}`),
-    ].join('\n');
-
-    return {
-      type: 'merge',
-      clusterIds: [...clusterIds],
-      mergedTitle,
-      mergedTags: allTags,
-      mergedLinks: allLinks,
-      mergedBody,
-      scope,
-    };
-  }
-
   /** Execute a merge — create consolidated engram and archive originals. */
   private executeMerge(cluster: Engram[], plan: ConsolidatePayload): void {
-    const dominantType = this.findDominantType(cluster);
+    const dominantType = findDominantType(cluster);
     const merged = this.engramService.create({
       title: plan.mergedTitle,
       body: plan.mergedBody,
@@ -226,21 +143,5 @@ export class ConsolidatorWorker extends WorkerBase {
     for (const original of cluster) {
       this.engramService.archive(original.id);
     }
-  }
-
-  private findDominantType(cluster: Engram[]): string {
-    const typeCounts = new Map<string, number>();
-    for (const e of cluster) {
-      typeCounts.set(e.type, (typeCounts.get(e.type) ?? 0) + 1);
-    }
-    let dominant = cluster[0]?.type ?? 'note';
-    let maxCount = 0;
-    for (const [type, cnt] of typeCounts) {
-      if (cnt > maxCount) {
-        dominant = type;
-        maxCount = cnt;
-      }
-    }
-    return dominant;
   }
 }
